+++ src/components/pages/Explore/Explore.tsx
@@ -7,8 +7,8 @@ import {PageStateEnum, FiltersConfig} from "../../../types";
 import MapboxMap from '../../map/MapboxMap';
 import MobileComponents from './ExploreMobile';
 import MaintenanceModal from "../../shared/MaintenanceModal";
-import LeftSidebar from "../../sidebar/left-sidebar/LeftSidebar";
-import Filters from "../../sidebar/right-sidebar/Filters";
+import LeftSidebar from "./left-sidebar/LeftSidebar";
+import Filters from "./right-sidebar/Filters";
 import Legend from "components/map/Legend";
 import { fetchExploreData } from "../../../utils/stateUpdateUtils";
 import { sendFiltersAnalyticsEvent, sendUnityAnalyticsEvent } from "../../../utils/analyticsUtils";
