+++ src/components/pages/Explore/right-sidebar/Filters.css
