+++ src/components/pages/Explore/right-sidebar/datepicker/DateSlider.css
