+++ src/components/pages/Explore/sidebar.scss
