+++ src/components/shared/LanguageSelector.tsx
@@ -1,8 +1,8 @@
 import React, { useContext, useState } from "react";
 import { CheckboxProps, Radio } from "semantic-ui-react";
-import { AppContext } from "../../../context";
-import { LanguageType } from "../../../types";
-import Translate from "../../../utils/translate/translateService";
+import { AppContext } from "../../context";
+import { LanguageType } from "../../types";
+import Translate from "../../utils/translate/translateService";
 
 
 export default function LanguageSelector() {
